Fixed JDK>8 stdlib source parsing
Turns out passing in an "empty" sourcefilemanager keeps the validator happy, which feels like a weird hack
But it seems to work? And we dont really need a full file manager as we're only parsing, not analyzing.
On that note, we're ditching analyzing. Even if the cost could be offset to making this whole spiel async,
the failure rate was too high (to be investigated), and with some more manual work of equality checking,
we should be able to get the same results for significantly cheaper and more "reliably"