Prioritize Maven packages over the JDK when exporting monikers
Previously, when a Maven package re-defined a class from the JDK then
lsif-java would emit an export against the JDK instead of the Maven
package. This caused problems with navigation to the JDK on Sourcegraph
because there were multiple repositories that exported monikers against
the `jdk/8` packageInformation node. Now, only the JDK repo should
export monikers for the JDK repo.