Upgrade to the latest sbt
The diff is unusually large because I have resisted to start using the
slash-syntax that was introduced ~2 years. The sbt v1.5 release
deprecated the old syntax that I've been using so now it's time to get
used to the new syntax. Fortunately, there's an automated migration
script so it was easy to create this PR