Swap order of signature and docstrings in hover messages
Previously, the docstring appeared above the signature in hover
messages. This was not ideal because it meant you had to scroll down to
see the signature when the docstring is very long (which happens
    frequently). This commit moves the signature to the top of
the hover message so you can always see it without scrolling.

Fixes https://github.com/sourcegraph/lsif-java/issues/284