Properly fix issue with checkerframework Gradle plugin.
Previously, `lsif-java index` had custom logic to disable the
checkerframework Gradle plugin in order to prevent the error
"plug-in not found: semanticdb". This commit replaces that custom logic
with a proper fix to this bug, which is to automatically inject the
SemanticDB compiler plugin jar to the annotation processor path
as well as the compile classpath.