Index SemanticDB synthetics
Previously, lsif-java only indexed the "Occurrences" section of a
SemanticDB `TextDocument`. Now, we additionall index the "Synthetics"
section, which includes inferred expression from Scala implicits, for
comprehensions, `.apply` methods and more. The benefit of this fix is
that "find references" on an implicit or an `apply` method now shows
usages where those symbols got inferred (the user didn't explicitly
write a reference to them).

It could be that this fix alone is not sufficient to enable "find
references" to show results on Sourcegraph. The Sourcegraph backend
will also need to accept offset ranges with zero length.