Add support for Java 17
Previously, the SemanticDB compiler plugin crashed when compiling code
with Java 17 because it was using private APIs. This commit refactors
the code to use blessed public APIs so that compiling with Java 17 works
successfully.

This commit only fixes the issues for the compiler plugin. Running
`lsif-java index` on Gradle/Maven projects will still fail on Java 17
projects since we don't automatically configure the requried
`--add-exports` flags.