Add support for Protobuf in Gradle projects
Previously, lsif-java only indexed the generated code from proto files.
Now, we also index the original proto files so that it's possible to
navigate from Java/Scala/Kotlin to Protobuf (and vice-versa).

This new feature only works for Gradle at the moment but the underlying
approach is generic and can be reused for other build tools as well.