Automate Docker container build and fix Java version detection
Previously, the Docker release process was manual. Now, the Docker
container is built by sbt on CI release and pushed to Docker Hub.

Also, the Docker container previously used the default Java version to
launch `lsif-java`. This caused problems in cases where the sources of
the project would be compiled with a different Java version from the
version that's specified in `dump.lsif`. Now, the LSIF version should
always match the version that was used to compile the sources.