Fix bug when indexing Scala package repo
Previously, we tried to infer the Scala version from the name of the jar
of the dependency that we were indexing. Now, we infer from the first
jar on the classpath that matches the `2_VERSION` pattern. The old logic
didn't work for the sbt package repo, which doesn't include the `_2.12`
suffix.