Improve support for Bazel builds, towards #88
This commit adds a new Bazel build to this codebase so that Bazel users can
build the semanticdb-javac compiler plugin from source and use it in
their Bazel build. This commit includes examples of how a Bazel user
could potentially enable the compiler plugin dynamically based on a
command-line flag, but we don't know if this is the best way to
integrate lsif-java with a Bazel codebase.

This commit is just an incremental step towards Bazel support, it's not
a complete solution yet because the `lsif-java index-semanticdb` command
still needs additional functionality to read jar files from the
`bazel-bin` directory.