Make the build work on Apple M1 chips
Previously, the build failed to load on M1 computers because Coursier
failed to fetch Java home for JVM version 8. Now, we force Coursier to
use amd64 instead of aarch64. This may result in slightly slower builds,
but it's a big improvement by the fact that the build loads and works
correctly now.