chore(ci): align usage of setup-java
I saw in one place this was using setup-java but the rest setup-scala.
Another opinionated change, but since setup-java also has the default
sbt launcher and built-in sbt caching I think it's a better choice to
just use that unless it's being included for another reason.

This also makes a small change on the sourcegraph upload to use the npm
installed src instead of curling it down manually.