dep: bump scalapb related stuff
This is an attempt to try to better align the protoc related tooling
since we've bumped the version of protobuf in the main build, but also
have older transitive stuff that is being brought in by these. Probably
best to bump them as well.

Sort of an attempt to see what might be going on with
https://github.com/sourcegraph/scip-java/issues/486.