fix: don't choke on scala/Function.
This took me an embarrassingly long time to catch what was going on ha.
But in the `SignatureFormatter` we special case things like `Function1`
to ensure we don't print `(A) => B` and instead `A => B`. We do this by
looking for symbols that start with `scala/Function`. However, the Scala
codebase has a `scala/Function.` object that causes this logic to trip
up. This change introduces another check to ensure that when we are in
Scala and see a symbol starting with `scala/Function` that it's not
`scala/Funtion.`.

This should get indexing working on the Scala codebase again.

Refs: #475