chore(build): skip source generation in import for Jmh
This change turns `bspEnabled := false` for `Jmh` because it involves
source generation meaning that when you do a build import it will also
do the source generation which we don't want on import.

NOTE: That if there is another release of sbt-jmh ever this is done by
default not, but not yet released: https://github.com/sbt/sbt-jmh/pull/207.