feat: account for -build-tool:mill
Without this if you're using Mill and you don't pass in -build-tool:sbt
this will blow up because it doesn't recognize vf: as a uri scheme when
default is choses. This introduces a new UriScheme.ZINC that both Mill
and sbt can use since as far as I know, this is a Zinc abstraction not
an sbt one. I then deprecated the UriScheme.SBT