Add `is_reference` relationships between synthetic case class symbols
Previously, doing "Find references" on a synthetic case class symbol
(for example, named parameter of a `copy` method) then it only showed
usages of that symbol and the non-synthetic symbol (for example, the
    case class field that matches the `copy` parameter). It didn't
return usages of other synthetic symbols that also relate to the case
class field (for example, the `apply` parameter).

This PR fixes the issue by adding `is_reference` relationships between
all synthetic symbols that relate to the same non-synthetic symbol.