Fix a regression where the SCIP index contained invalid symbols
Uploading an index with scip-java v0.8.17 results in the following
error in the Sourcegraph UI

```
codeNavSvc.GetImplementations: lsifStore.MonikersByPosition: reached end of symbol while parsing <scheme>, expected a ' ' character
minimized/Issue413Subclass#c.
_____________________________^
```
The problem is that this release introduced a regression where scip-java
emitted invalid SCIP symbols. This PR fixes the issue, and adds new
checks in the testing infrastructure to prevent this kind of regression
from happening again.