Bazel: add support for DirectoryFileObject
Previously, the scip-java Bazel support only supported SimpleFileObject
in the Java compiler API and it crashed when encountering other types of
file objects like DirectoryFileObject, which are used in some Bazel
builds in the wild. This commit adds support for DirectoryFileObject.

Additionally, this commit adds more detailed documentation on how to
use scip-java with Bazel.