Add option --allow-exporting-global-symbols-from-directory-entries
Previously, scip-java only allowed cross-index (or cross-repo)
navigation between symbols that are compiled to classfiles inside jar
files. This behavior mixed an intentional design decision with an
accidental implementation detail. The intentional design decision was
that the ideal cross-repo navigation experience is from repositories to
packages. To achieve this behavior, scip-java didn't assign a package
name and package version to symbols that are compiled to directory
entries on the classpath.  The accidental implementation detail was that
Gradle, Maven, and sbt compile to directories while packages only have
jar files.

This commit changes the behavior to treat symbols the same regardless if
they are compiled to jar files or directories. The accidental
implemenation detail was a hack, and this behavior has come up as a
surprise in at least twice cases of trying to integrate scip-java with
big propriatery codebases. To get the default old behavior, the flag
`--allow-exporting-global-symbols-from-directory-entries` can be set to
false.