Fix "Find references" bug for interfaces and types
Fixes #678

Previously, doing "Find references" on a symbol with kind `INTERFACE` or
`TYPE` returned results for all references to symbols that *implement*
that symbol. This bug was particularly noticeable for Kotlin sources because
scip-kotlin emits the `TYPE` kind for classes. Now, we no longer emit
reference relationships so the implementations only appear in "Find
implementation" and not "Find references".